Show key bindings in completions
This feature is currently rudimentary but functional. It maintains a
hash table of known key bindings based on all the keymaps in
"(current-active-maps)". This is much faster than doing a key binding
lookup on every command in the obarray. It's fast enough to run on
every invocation of smex, although it mostly only needs to run after
switching buffers (which changes the set of currently-active kye
bindings based on active modes). It would benefit from some idle
timers to update the keybind hash in the background when the user
changes buffers.