Consider command without key binding as a completion match
When "smex-show-key-bindings" was non-nil and using standard Emacs
completion, if a command name was complete but not unique (e.g.
"find-file"), then it was required to type in the keyboard shortcut as
well, since the command by itself was no longer considered to be
complete. That has been fixed by using "completion-table-in-turn" to
also consider the command by itself as a complete match.

A test is also added to verify the fix.